Issue #202 - Move work to a new controller
